refactor: create `backtesting.models`.
